created new file for animation
